# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  color: string;
  createdAt: string;
  updatedAt: string;
}

export type ProjectWithStatsMyRole =
  (typeof ProjectWithStatsMyRole)[keyof typeof ProjectWithStatsMyRole];

export const ProjectWithStatsMyRole = {
  admin: "admin",
  member: "member",
} as const;

export type ProjectWithStats = Project & {
  totalTasks: number;
  completedTasks: number;
  overdueTasks: number;
  memberCount: number;
  myRole: ProjectWithStatsMyRole;
};

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
  color?: string;
}

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  color?: string;
}

export type ProjectMemberRole =
  (typeof ProjectMemberRole)[keyof typeof ProjectMemberRole];

export const ProjectMemberRole = {
  admin: "admin",
  member: "member",
} as const;

export interface ProjectMember {
  id: number;
  projectId: number;
  userId: number;
  role: ProjectMemberRole;
  joinedAt: string;
  user: User;
}

export type AddProjectMemberBodyRole =
  (typeof AddProjectMemberBodyRole)[keyof typeof AddProjectMemberBodyRole];

export const AddProjectMemberBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface AddProjectMemberBody {
  email: string;
  role: AddProjectMemberBodyRole;
}

export type UpdateMemberRoleBodyRole =
  (typeof UpdateMemberRoleBodyRole)[keyof typeof UpdateMemberRoleBodyRole];

export const UpdateMemberRoleBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface UpdateMemberRoleBody {
  role: UpdateMemberRoleBodyRole;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
  cancelled: "cancelled",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface Task {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  assigneeId: number | null;
  /** @nullable */
  dueDate: string | null;
  createdById: number;
  createdAt: string;
  updatedAt: string;
}

export type TaskWithAssignee = Task & {
  assignee: User | null;
  createdBy: User;
};

export type TaskWithProject = Task & {
  project: Project;
  assignee: User | null;
};

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
  cancelled: "cancelled",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status?: CreateTaskBodyStatus;
  priority?: CreateTaskBodyPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
  cancelled: "cancelled",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface PriorityCount {
  priority: string;
  count: number;
}

export interface DashboardSummary {
  totalProjects: number;
  totalTasks: number;
  myOpenTasks: number;
  overdueTasks: number;
  completedThisWeek: number;
  tasksByStatus: StatusCount[];
  tasksByPriority: PriorityCount[];
}

export interface ActivityItem {
  id: number;
  taskId: number;
  taskTitle: string;
  projectId: number;
  projectName: string;
  action: string;
  actorName: string;
  /** @nullable */
  actorAvatarUrl: string | null;
  createdAt: string;
}

export type ListTasksParams = {
  status?: ListTasksStatus;
  assigneeId?: number;
  priority?: ListTasksPriority;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
  cancelled: "cancelled",
} as const;

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;